+++ package.json
@@ -19,6 +19,7 @@
     "react-native-reanimated": "^1.13.2",
     "react-native-safe-area-context": "^3.1.9",
     "react-native-screens": "^2.16.1",
+    "react-native-splash-screen": "^3.2.0",
     "react-native-video": "^5.1.0-alpha8",
     "typescript": "^4.1.3"
   },
