+++ src/App.js
@@ -10,9 +10,9 @@ import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import Main from './screens/Main';
-import Show from './screens/Show';
+import ShowScreen from './screens/ShowScreen';
 import EpisodeDetails from './screens/VideoPlayer';
-import ShowList from './screens/ShowList';
+import ShowList from './screens/ShowListScreen';
 
 const Stack = createStackNavigator();
 
@@ -20,14 +20,14 @@ const App: () => React$Node = () => {
     return (
         <NavigationContainer>
             <Stack.Navigator>
+                <Stack.Screen name="ShowLists" component={ShowList}
+                              options={{headerShown: false}} />
                 <Stack.Screen
                     name="Home"
                     component={Main}
                     options={{headerShown: false}}
                 />
-                <Stack.Screen name="ShowLists" component={ShowList}
-                              options={{headerShown: false}} />
-                <Stack.Screen name="Details" component={Show}
+                <Stack.Screen name="Details" component={ShowScreen}
                               options={{headerShown: false}} />
                 <Stack.Screen name="EpisodeDetails" component={EpisodeDetails}
                               options={{headerShown: false}}/>
