+++ src/screens/Main.js
@@ -7,8 +7,8 @@
  */
 
 import React from 'react';
-import {SafeAreaView, ScrollView, StatusBar, StyleSheet, Text, View} from 'react-native';
-
+import {SafeAreaView, ScrollView, StyleSheet, Text, View} from 'react-native';
+import SplashScreen from 'react-native-splash-screen';
 import {Colors} from 'react-native/Libraries/NewAppScreen';
 import Header from '../components/Header';
